+++ test/test-helpers/setup-repository.ts
@@ -1,5 +1,4 @@
 import * as fs from 'fs';
-import * as mkdirp from 'mkdirp';
 import * as os from 'os';
 import * as path from 'path';
 import { promisify } from 'util';
@@ -7,13 +6,14 @@ import { v4 as guid } from 'uuid';
 
 import { Git } from '../../src/git/Git';
 
+const mkdir = promisify(fs.mkdir);
 const writeFile = promisify(fs.writeFile);
 
-export async function setupRepository(): Promise<string> {
-    let root: string;
-
-    root = path.join(os.tmpdir(), guid());
-    await mkdirp(root);
+export async function setupRepository(root?: string): Promise<string> {
+    if (!root) {
+        root = path.join(os.tmpdir(), guid());
+        await mkdir(root, { recursive: true });
+    }
 
     await Git.execute(root, 'init');
     await Git.execute(root, 'config', 'user.email', 'foo@example.com');